+++ collector/index.js
@@ -8,7 +8,7 @@ const bodyParser = require("body-parser");
 const cors = require("cors");
 const path = require("path");
 const { ACCEPTED_MIMES } = require("./utils/constants");
-const { reqBody } = require("./utils/http");
+const { reqBody, getCollectorPort } = require("./utils/http");
 const { processSingleFile } = require("./processSingleFile");
 const { processLink, getLinkText } = require("./processLink");
 const { wipeCollectorStorage } = require("./utils/files");
@@ -18,18 +18,7 @@ const { verifyPayloadIntegrity } = require("./middleware/verifyIntegrity");
 const { httpLogger } = require("./middleware/httpLogger");
 const app = express();
 const FILE_LIMIT = "3GB";
-const DEFAULT_COLLECTOR_PORT = 8888;
 const COLLECTOR_PORT = getCollectorPort();
-
-function getCollectorPort() {
-  const port = Number(process.env.COLLECTOR_PORT || DEFAULT_COLLECTOR_PORT);
-  if (Number.isInteger(port) && port > 0 && port <= 65535) return port;
-
-  console.warn(
-    `Invalid COLLECTOR_PORT "${process.env.COLLECTOR_PORT}". Falling back to ${DEFAULT_COLLECTOR_PORT}.`
-  );
-  return DEFAULT_COLLECTOR_PORT;
-}
 
 // Only log HTTP requests in development mode and if the ENABLE_HTTP_LOGGER environment variable is set to true
 if (
