+++ collector/utils/http/index.js
@@ -1,6 +1,7 @@
 process.env.NODE_ENV === "development"
   ? require("dotenv").config({ path: `.env.${process.env.NODE_ENV}` })
   : require("dotenv").config();
+const DEFAULT_COLLECTOR_PORT = 8888;
 
 function reqBody(request) {
   return typeof request.body === "string"
@@ -28,8 +29,26 @@ function validBaseUrl(baseUrl) {
   }
 }
 
+/**
+ * Gets the collector port from the environment variables.
+ * If the port is not set, it will fall back to the default port.
+ * If the port is invalid, it will log a warning and return the default port.
+ * @returns {number}
+ */
+function getCollectorPort() {
+  if (!("COLLECTOR_PORT" in process.env)) return DEFAULT_COLLECTOR_PORT;
+  const port = Number(process.env.COLLECTOR_PORT);
+  if (Number.isInteger(port) && port > 0 && port <= 65535) return port;
+
+  console.warn(
+    `Invalid COLLECTOR_PORT "${process.env.COLLECTOR_PORT}". Falling back to ${DEFAULT_COLLECTOR_PORT}.`
+  );
+  return DEFAULT_COLLECTOR_PORT;
+}
+
 module.exports = {
   reqBody,
   queryParams,
   validBaseUrl,
+  getCollectorPort,
 };
